use db directly in components; allow index on $deleted and $synchronized
closes #11, closes #12

+++ demo-app/schema.sql
@@ -9,7 +9,7 @@ CREATE TABLE `tasks` (
   -- Required columns per table
   `id` VARCHAR(36) NOT NULL PRIMARY KEY,
   `$updated` BIGINT(14) NOT NULL DEFAULT 0,
-  `$deleted` TINYINT(1) NOT NULL DEFAULT 0,
+  `$deleted` INTEGER(1) NOT NULL DEFAULT 0,
   `$synchronized` BIGINT(14) NOT NULL DEFAULT 0,
 
   -- Optional customized columns per table
@@ -23,7 +23,7 @@ CREATE TABLE `files` (
   -- Required columns per table
   `id` VARCHAR(36) NOT NULL PRIMARY KEY,
   `$updated` BIGINT(14) NOT NULL DEFAULT 0,
-  `$deleted` TINYINT(1) NOT NULL DEFAULT 0,
+  `$deleted` INTEGER(1) NOT NULL DEFAULT 0,
   `$synchronized` BIGINT(14) NOT NULL DEFAULT 0,
 
   -- Optional customized columns per table
